Rename server creation module
